+++ tests/Unit/Traits/DomainTestingTrait.php
@@ -23,6 +23,7 @@ declare(strict_types=1);
 namespace OAT\Library\Lti1p3Ags\Tests\Unit\Traits;
 
 use DateTimeInterface;
+use OAT\Library\Lti1p3Ags\Factory\ScoreFactory;
 use OAT\Library\Lti1p3Ags\Model\Score;
 use OAT\Library\Lti1p3Core\Platform\Platform;
 use OAT\Library\Lti1p3Core\Platform\PlatformInterface;
@@ -87,10 +88,10 @@ trait DomainTestingTrait
         ?float $scoreMaximum = 0.3,
         ?string $comment = null,
         ?DateTimeInterface $timestamp = null,
-        ?string $activityProgressStatus = null,
-        ?string $gradingProgressStatus = null
+        string $activityProgressStatus = Score::ACTIVITY_PROGRESS_STATUS_INITIALIZED,
+        string $gradingProgressStatus = Score::GRADING_PROGRESS_STATUS_NOT_READY
     ): Score {
-        return new Score(
+        return (new ScoreFactory())->create(
             $userId,
             $contextId,
             $lineItemId,