+++ CHANGELOG.md
@@ -5,7 +5,7 @@ unreleased
 -----
 
 * Added ScoreFactory
-* Modified score id attribute to identifier
+* Modified Score id attribute to identifier
 
 1.0.2
 -----